requirements.txt
anyio==4.5.2
APScheduler==3.6.3
backports.zoneinfo==0.2.1
certifi==2024.12.14
cffi==1.17.1
cryptography==44.0.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pycparser==2.22
python-dotenv==1.0.1
python-telegram-bot==13.2
pytimeparse==1.1.8
pytz==2024.2
six==1.17.0
sniffio==1.3.1
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2

